Adapt to new structure of links data
Following change in https://github.com/w3c/reffy/pull/1055
Also split code between CLI and re-usable library
